refactor: move and extract SearchResults,NoSearchResults,SearchResultsItem,SearchResultsSubjectIcon
