--- src/components/SearchResults/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./SearchResults";
