--- src/components/SearchResultsItem/index.tsx
@@ -1,1 +1,0 @@
-export { default } from "./SearchResultsItem";
