+++ src/components/TeacherComponents/NoSearchResults/NoSearchResults.test.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 
-import renderWithProviders from "../../__tests__/__helpers__/renderWithProviders";
-
 import NoSearchResults from "./NoSearchResults";
+
+import renderWithProviders from "@/__tests__/__helpers__/renderWithProviders";
 
 const render = renderWithProviders();
