+++ src/components/TeacherComponents/NoSearchResults/NoSearchResults.tsx
