+++ src/components/TeacherComponents/NoSearchResults/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./NoSearchResults";
