+++ src/components/TeacherComponents/SearchResults/SearchResults.stories.tsx
@@ -1,12 +1,11 @@
 import React from "react";
 import { ComponentStory, ComponentMeta } from "@storybook/react";
 
-import AnalyticsDecorator from "../../storybook-decorators/AnalyticsDecorator";
-import elasticResponseFixture from "../../context/Search/elasticResponse.2020.fixture.json";
-import teachersHomePageFixture from "../../node-lib/curriculum-api/fixtures/teachersHomePage.fixture";
-
 import Component from "./SearchResults";
 
+import AnalyticsDecorator from "@/storybook-decorators/AnalyticsDecorator";
+import elasticResponseFixture from "@/context/Search/elasticResponse.2020.fixture.json";
+import teachersHomePageFixture from "@/node-lib/curriculum-api/fixtures/teachersHomePage.fixture";
 import { searchResultsHitsSchema } from "@/context/Search/search.schema";
 
 export default {
