+++ src/components/TeacherComponents/SearchResults/SearchResults.tsx
