+++ src/components/TeacherComponents/SearchResults/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./SearchResults";
+export { RESULTS_PER_PAGE } from "./SearchResults";
