+++ src/components/TeacherComponents/SearchResultsItem/SearchResultsItem.stories.tsx
@@ -1,12 +1,11 @@
 import React from "react";
 import { ComponentStory, ComponentMeta } from "@storybook/react";
 
-import AnalyticsDecorator from "../../storybook-decorators/AnalyticsDecorator";
-import elasticResponseFixture from "../../context/Search/elasticResponse.2020.fixture.json";
-import teachersHomePageFixture from "../../node-lib/curriculum-api/fixtures/teachersHomePage.fixture";
-
 import Component from "./SearchResultsItem";
 
+import AnalyticsDecorator from "@/storybook-decorators/AnalyticsDecorator";
+import elasticResponseFixture from "@/context/Search/elasticResponse.2020.fixture.json";
+import teachersHomePageFixture from "@/node-lib/curriculum-api/fixtures/teachersHomePage.fixture";
 import { searchResultsHitSchema } from "@/context/Search/search.schema";
 import { getSearchHitObject } from "@/context/Search/search.helpers";
