+++ src/components/TeacherComponents/SearchResultsItem/index.tsx
@@ -1,0 +1,2 @@
+export { default } from "./SearchResultsItem";
+export type { SearchResultsItemProps } from "./SearchResultsItem";
