+++ src/components/TeacherComponents/SearchResultsSubjectIcon/SearchResultsSubjectIcon.tsx
