+++ src/components/TeacherComponents/SearchResultsSubjectIcon/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./SearchResultsSubjectIcon";
