+++ src/components/TeacherViews/Search/helpers/index.ts
