+++ src/pages/teachers/search.tsx
@@ -3,7 +3,7 @@ import { GetStaticProps, NextPage } from "next";
 
 import AppLayout from "@/components/AppLayout";
 import useSearch from "@/context/Search/useSearch";
-import Search from "@/components/TeacherViews/SearchComponents/Search.view";
+import Search from "@/components/TeacherViews/Search/Search.view";
 import curriculumApi2023, {
   SearchPageData,
 } from "@/node-lib/curriculum-api-2023";
@@ -11,7 +11,7 @@ import curriculumApi from "@/node-lib/curriculum-api";
 import { getSeoProps } from "@/browser-lib/seo/getSeoProps";
 import useSearchFilters from "@/context/Search/useSearchFilters";
 import usePagination from "@/components/SharedComponents/Pagination/usePagination";
-import { RESULTS_PER_PAGE } from "@/components/SearchResults/SearchResults";
+import { RESULTS_PER_PAGE } from "@/components/TeacherComponents/SearchResults";
 import getPageProps from "@/node-lib/getPageProps";
 
 type SearchPageProps = {
